Pref Source Maps with Source Maps (#879)
* Add shouldSourceMap util
+ add prefs utility for tracking debugger prefs
+ add prefsHelper for cleaning up utils
* Remove defaultPrefsEnabled

+++ bin/import-deps.js
@@ -21,6 +21,11 @@ glob("public/js/lib/devtools/**/*.js").forEach((debuggerFile) => {
     path.relative("public/js/lib/", debuggerFile)
   );
 
-  const fileText = fs.readFileSync(geckoFilePath, 'utf8');
-  fs.writeFileSync(debuggerFile, fileText);
+
+  if (fs.existsSync(geckoFilePath)) {
+    const fileText = fs.readFileSync(geckoFilePath, 'utf8');
+    fs.writeFileSync(debuggerFile, fileText);
+  } else {
+    console.log(`file: ${geckoFilePath} does not exist`);
+  }
 })
